Moved get_preparations from admin to sample routes
